Backend/requirements.txt
fastapi
uvicorn
ultralytics
python-multipart
pillow
numpy
torch
torchvision
opencv-python-headless
